Portability changes.

+++ alLib.c
@@ -57,7 +57,7 @@ char *Strncat(
   /* max must be >= 0 and no more than stringsize - 1 */
   /* for char string[10];       max must be <= 9 */
 
-int l, newMax;
+size_t l, newMax;
 char *s;
 
   l = strlen( dest );
